add OwnerPolicy authorization

+++ src/datingApp.Application/Commands/Handlers/DeleteMatchHandler.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using datingApp.Application.Abstractions;
 using datingApp.Application.Exceptions;
 using datingApp.Application.Repositories;
+using datingApp.Application.Security;
 using datingApp.Core.Repositories;
 
 namespace datingApp.Application.Commands.Handlers;
@@ -14,10 +15,13 @@ public sealed class DeleteMatchHandler : ICommandHandler<DeleteMatch>
 {
     private readonly IMatchRepository _matchRepository;
     private readonly IDeletedEntityRepository _deletedEntityRepository;
-    public DeleteMatchHandler(IMatchRepository matchRepository, IDeletedEntityRepository deletedEntityRepository)
+    private readonly IDatingAppAuthorizationService _authorizationService;
+
+    public DeleteMatchHandler(IMatchRepository matchRepository, IDeletedEntityRepository deletedEntityRepository, IDatingAppAuthorizationService authorizationService)
     {
         _matchRepository = matchRepository;
         _deletedEntityRepository = deletedEntityRepository;
+        _authorizationService = authorizationService;
     }
     public async Task HandleAsync(DeleteMatch command)
     {
@@ -33,7 +37,13 @@ public sealed class DeleteMatchHandler : ICommandHandler<DeleteMatch>
                 throw new MatchNotExistsException(command.MatchId);
             }
         }
-        
+
+        var authorizationResult = await _authorizationService.AuthorizeAsync(command.AuthenticatedUserId, match, "OwnerPolicy");
+        if (!authorizationResult.Succeeded)
+        {
+            throw new UnauthorizedException();
+        }
+
         await _matchRepository.DeleteAsync(match);
         await _deletedEntityRepository.AddAsync(match.Id);
     }
